modifying arg of function

+++ 0x14-bit_manipulation/1-print_binary.c
@@ -6,7 +6,7 @@
  * @n: the decimal number.
  */
 
-void print_binary(unsigned int n)
+void print_binary(unsigned long int n)
 {
 	unsigned long int a;
 	int b;
